refactor: изменить название файла в папке about-us page.tsx на index.tsx
